HTML documentation to MarkDown documentation
The original documentation is disappearing from Internet, along with the author's website.
This is the manual conversion of the original documentation from HTML to MarkDown, so it will be available in the future along with the utilities. The text has been manipulated so as to eliminate unnecessary external references and appear correctly on GitHub and similar. The sarcasm of the author has been preserved, as it should be. :-)